requirements.txt
pandas>=1.3.5
python-dotenv>=0.21.0
pydantic>=1.10.0,<2.0.0
pytest>=7.4.0
tenacity>=8.2.0
requests>=2.28.0
pyyaml>=6.0
